Move comment menu items into their own folder
