+++ packages/lesswrong/lib/components.ts
@@ -244,27 +244,27 @@ if (getSetting('hasEvents', true)) {
 
 importComponent("CommentsItem", () => require('../components/comments/CommentsItem/CommentsItem'));
 importComponent("CommentUserName", () => require('../components/comments/CommentsItem/CommentUserName'));
-importComponent("RetractCommentMenuItem", () => require('../components/comments/CommentsItem/RetractCommentMenuItem'));
-importComponent("MoveToAnswersMenuItem", () => require('../components/comments/CommentsItem/MoveToAnswersMenuItem'));
-importComponent("CommentsPermalinkMenuItem", () => require('../components/comments/CommentsItem/CommentsPermalinkMenuItem'));
+importComponent("RetractCommentMenuItem", () => require('../components/comments/CommentActions/RetractCommentMenuItem'));
+importComponent("MoveToAnswersMenuItem", () => require('../components/comments/CommentActions/MoveToAnswersMenuItem'));
+importComponent("CommentsPermalinkMenuItem", () => require('../components/comments/CommentActions/CommentsPermalinkMenuItem'));
 importComponent("CommentShortformIcon", () => require('../components/comments/CommentsItem/CommentShortformIcon'));
-importComponent("BanUserFromPostMenuItem", () => require('../components/comments/CommentsItem/BanUserFromPostMenuItem'));
-importComponent("BanUserFromAllPostsMenuItem", () => require('../components/comments/CommentsItem/BanUserFromAllPostsMenuItem'));
-importComponent("BanUserFromAllPersonalPostsMenuItem", () => require('../components/comments/CommentsItem/BanUserFromAllPersonalPostsMenuItem'));
-importComponent("DeleteCommentMenuItem", () => require('../components/comments/CommentsItem/DeleteCommentMenuItem'));
+importComponent("BanUserFromPostMenuItem", () => require('../components/comments/CommentActions/BanUserFromPostMenuItem'));
+importComponent("BanUserFromAllPostsMenuItem", () => require('../components/comments/CommentActions/BanUserFromAllPostsMenuItem'));
+importComponent("BanUserFromAllPersonalPostsMenuItem", () => require('../components/comments/CommentActions/BanUserFromAllPersonalPostsMenuItem'));
+importComponent("DeleteCommentMenuItem", () => require('../components/comments/CommentActions/DeleteCommentMenuItem'));
 importComponent("DeleteCommentDialog", () => require('../components/comments/CommentsItem/DeleteCommentDialog'));
-importComponent("EditCommentMenuItem", () => require('../components/comments/CommentsItem/EditCommentMenuItem'));
-importComponent("ReportCommentMenuItem", () => require('../components/comments/CommentsItem/ReportCommentMenuItem'));
-importComponent("MoveToAlignmentMenuItem", () => require('../components/comments/CommentsItem/MoveToAlignmentMenuItem'));
-importComponent("SuggestAlignmentMenuItem", () => require('../components/comments/CommentsItem/SuggestAlignmentMenuItem'));
-importComponent("SubscribeToCommentMenuItem", () => require('../components/comments/CommentsItem/SubscribeToCommentMenuItem'));
+importComponent("EditCommentMenuItem", () => require('../components/comments/CommentActions/EditCommentMenuItem'));
+importComponent("ReportCommentMenuItem", () => require('../components/comments/CommentActions/ReportCommentMenuItem'));
+importComponent("MoveToAlignmentMenuItem", () => require('../components/comments/CommentActions/MoveToAlignmentMenuItem'));
+importComponent("SuggestAlignmentMenuItem", () => require('../components/comments/CommentActions/SuggestAlignmentMenuItem'));
+importComponent("SubscribeToCommentMenuItem", () => require('../components/comments/CommentActions/SubscribeToCommentMenuItem'));
 importComponent("CommentDeletedMetadata", () => require('../components/comments/CommentsItem/CommentDeletedMetadata'));
 importComponent("CommentBody", () => require('../components/comments/CommentsItem/CommentBody'));
-importComponent("CommentActions", () => require('../components/comments/CommentsItem/CommentActions'));
+importComponent("CommentActions", () => require('../components/comments/CommentActions/CommentActions'));
 importComponent("CommentsMenu", () => require('../components/comments/CommentsItem/CommentsMenu'));
 importComponent("CommentOutdatedWarning", () => require('../components/comments/CommentsItem/CommentOutdatedWarning'));
 importComponent("CommentsItemDate", () => require('../components/comments/CommentsItem/CommentsItemDate'));
-importComponent("ToggleIsModeratorComment", () => require('../components/comments/CommentsItem/ToggleIsModeratorComment'));
+importComponent("ToggleIsModeratorComment", () => require('../components/comments/CommentActions/ToggleIsModeratorComment'));
 
 importComponent("AllComments", () => require('../components/comments/AllComments'));
